move-modeling_v3/scripts/requirements.txt
certifi==2021.10.8
dnspython==2.1.0
haversine==2.5.1
numpy==1.21.4
paho-mqtt==1.6.1
pygame==2.1.0
pyproj==3.3.0
python-etcd==0.4.5
Shapely==1.8.0
urllib3==1.26.7
